This commit was generated by cvs2svn to compensate for changes in r18907, which included commits to RCS files with non-trunk default branches.
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@18908 728af825-273c-0410-89f9-f7d3b574a069

+++ third/tetex/src/texk/dvipsk/dopage.c
@@ -112,8 +112,8 @@ dopage P1H(void)
    H_BREAK = (30 * DPI / 400 ) ; /* 30 seems to have worked well at 400 dpi */
 #endif 
    w = x = y = z = 0 ;
-   h = DPI / conv * hoff / 4736286.72 ;
-   v = DPI / conv * voff / 4736286.72 ;
+   h = (integer) (DPI / conv * hoff / 4736286.72) ;
+   v = (integer) (DPI / conv * voff / 4736286.72) ;
    hh = PixRound(h) ;
    vv = PixRound(v) ;
    curfnt = NULL ;
